Back -- Refactor -- C# Watch database changes

+++ nodes/hub/back/Db/Configs/DbConfig.cs
@@ -4,7 +4,7 @@
 	{
 		public const string Section = "Database";
 		public string Username { get; set; }
-		public long Port { get; set; }
+		public int Port { get; set; }
 		public string Database { get; set; }
 		public string Password { get; set; }
 		public string Host { get; set; }
